Updated a few things in timeslot/index.blade.php and scripts.blade.php
Updated the timeslot script where a bug happens that a jock is always been added
while adding timeslots to shows. Disabled the jock/show form depending on what the
user chooses, jock_id form is disabled when schedule type is show, and on the other
hand show_id form is disabled when schedule type is jock.